requirements.txt
pandas==1.4.4
numpy==1.21.5
scikit-learn==1.0.2
xgboost==1.6.1
mlflow==1.26.1
pyyaml==6.0
pytest==7.1.2
python-dotenv==0.20.0
matplotlib==3.5.2
seaborn==0.11.2
joblib==1.1.0 